Implemented the search parameter :not modifier for the Token type.
This fails to return resources where the Token is ':not' found at all (e.g. patient.gender is not found at all)!
To fix this I think we need to beginning to add rows in the index for each possible search parameter, even where there is no matched data in the resource, that is, add a row where the values are null.
Then here we can search for those rows where the ':not' Modifier is used, in fact, the NotEqualTo will just find them