Big refactor moving files around the project and updating namespaces. This was to achieve a refinement to the application architecture strategy. Following these rules:
Application project it to only hold services or classes of a service nature. Classes that are core to the applications' primary logic.

Domain project holds all entity type classes including Records, Enums, Request, Response or Outcome  classes. Yet it also hold support classes which do hold logic, yet this logic is more of a support nature and more generally used by the whole solution.

In general, and by default an Application project service class is to have its Interface also in the Application layer, unless that interface is required in the Domain layer, in which case it is moved there. In theory all Interfaces should be in the Domain layer, yet I feel this is too heavy-handed.